Bring the code from transport_ofi.impl.h directly into transport_ofi.h. Hopefully this will make merges easier.
